Fixed below bugs identified in the datepicker
Prev arrow is not visible properly
1. Set a startDate
2. and set the value to the startdate
3. When clicked on the input box, the prev arrow is not visible.
4. When moved to the next month, the prev arrow is still not visible.

Time picker for 12:00 AM/PM with the hh:mm tt format.
1. Set the format to "hh:mm tt" and value to "12:00 PM"/"12:00 AM" and enable only timepicker
2. Click on input and press tab, the value changes to "00:00 PM", instead of staying at the same value

Display does not update when the text box receives focus
1. Select a date to the picker
2. Click outside the picker and hide the selector.
3. Set the value of the input box to a new value using javascript
4. Click inside the date box. The calendar does not change to the new date selected.

Set the icon to time icon when only pickTime is enabled

Correct the date field to proper format on lost focus, when the datepicker is not visible.